added modify.go
