+++ fasta/info.go
@@ -1,42 +1,16 @@
 package fasta
 
 import (
-	"github.com/vertgenlab/gonomics/dna"
-	"log"
 	"strings"
 )
 
-func Extract(f *Fasta, start int64, end int64, name string) *Fasta {
-	var ans Fasta
-	if start < 0 || end < 0 || start > int64(len(f.Seq)) || end > int64(len(f.Seq)) || start > end {
-		log.Fatalf("Invalid start and end given to fasta.Extract. Start: %d, End: %d, ChromLength: %d.", start, end, len(f.Seq))
-	}
-	ans.Seq = dna.Extract(f.Seq, int(start), int(end))
-	ans.Name = name
-	return &ans
-}
-
-func CountBase(fa *Fasta, b dna.Base) int64 {
-	var answer int64 = 0
-	for i := 0; i < len(fa.Seq); i++ {
-		if fa.Seq[i] == b {
-			answer++
-		}
-	}
-	return answer
-}
-
-func FindFaIndex(subFa []*Fasta, n string) int {
-	for i := 0; i < len(subFa); i++ {
-		if subFa[i].Name == n {
-			return i
-		}
-	}
-	return -1
-}
-
+// IsFasta returns true if the input filename has a fasta file extension.
+// Input filename may have a .gz suffix.
 func IsFasta(filename string) bool {
-	if strings.HasSuffix(filename, ".fasta") || strings.HasSuffix(filename, ".fa") || strings.HasSuffix(filename, ".fasta.gz") || strings.HasSuffix(filename, ".fa.gz") {
+	if strings.HasSuffix(filename, ".fasta") ||
+		strings.HasSuffix(filename, ".fa") ||
+		strings.HasSuffix(filename, ".fasta.gz") ||
+		strings.HasSuffix(filename, ".fa.gz") {
 		return true
 	} else {
 		return false
