+++ fasta/randomReads.go
@@ -23,7 +23,7 @@ func RandomSeqGenerator(genome []*Fasta, readLength int, readNumber int) []*Fast
 		start = randIntInRange(0, len(genome[chrom].Seq)-readLength)
 		readName = genome[chrom].Name + "_" + strconv.Itoa(start) + "_" + strconv.Itoa(start+readLength)
 
-		if CountBase(genome[chrom], dna.N) == 0 {
+		if dna.CountBase(genome[chrom].Seq, dna.N) == 0 {
 			answer = append(answer, &Fasta{Name: readName, Seq: genome[chrom].Seq[start : start+readLength]})
 			i++
 		} else {
