+++ fasta/stream.go
