+++ giraf/binaryGiraf/testdata/readtest.giraf
@@ -1,0 +1,1 @@
+goldenState	0	99	0	+	50:1>2:50	100=	110335	5	ATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCG	kkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkfkkkkf	BZ:i:5	GP:Z:30,11,35,9,23	XO:i:5
